We gave up rewriting OpenGL in GX so we decided to use the gl2gx library instead. Everything compiled with minimal changes made to the original OGLRender files.
We then made a lot of progress with the port by linking in SDL Wii and getting a lot of more files to compile.

Files that are port is not going to use as well as files we decided to port a different way were moved into a new folder called unusedsource for reference if we needed to llok at them again down the road.