MAJOR CHANGE!
Sorry everyone, but this had to happen sooner or later. We need to be more organized with our files. I have altered the file structure to look in sd:/DS/ROMS/ for roms, for example, instead of "DSROM". It is my recommendation that you make a folder "SAVES" and "BIOS" in the DS folder, and keep everything related to the DS in there too. This will make everything a lot cleaner on our cards! I was going to name it "desmumeWii" instead of DS, but I didn't want to run the risk of hitting the maximum character limit (256). 

I also cleaned up main.cpp, and made it use hardware timers instead of what we had before.