Two words: Paired. Single.
That's right. I got tired of waiting. I created paired-single versions of ALL of the matrix functions. On average they run about 40-50% faster. This has been in the works for a long time. There isn't any noticeable speed up that I could find, but it's there, and a preview of things to come.